[Feat] UserVerify 프로토콜 기본 메서드 구현 Implements 레이어로 이동
