+++ Projects/Domain/OAuth/Sources/Service/AppleLogInService.swift
@@ -31,7 +31,6 @@ extension AppleLogInService: @retroactive OAuthServiceProtocol {
         for await userIdentityResult in identityStream {
             switch userIdentityResult {
             case .success(let userIdentity):
-                print("user Identity: \(userIdentity)")
                 return .apple
             case .failure(let failure): throw failure
             }
